User authentication and error handling

+++ modules/db.js
@@ -1,6 +1,12 @@
 const pg = require('pg');
 
-const db = function(dbConnectionString) {
+const DB_RESPONSES = {
+    OK: "OK",
+    NOT_EXIST: "NOT_EXIST",
+    ALREADY_EXIST: "ALREADY_EXIST"
+};
+
+const db = function (dbConnectionString) {
     const connectionString = dbConnectionString;
 
     async function runQuery(query, params) {
@@ -19,44 +25,94 @@ const db = function(dbConnectionString) {
         await client.end();
     }
 
-
-    
-    const getUserByID = async function(userID) {
+    const getUserByName = async function (userName) {
         let userData = null;
         try {
-            userData = await runQuery(`SELECT * FROM users WHERE userID=$1`, [userID]);
-        } catch(error) {
+            userData = await runQuery('SELECT * FROM users WHERE name=$1', [userName]);
+        } catch (error) {
             console.error(error);
         }
         return userData;
     }
 
-    const insertUser = async function(userName, userEmail, userPassword) {
+    const getUserByEmail = async function (userName) {
+        let userData = null;
         try {
-            await insertData(`INSERT INTO users(name, email, password) VALUES ($1, $2, $3)`, [userName, userEmail, userPassword]);
-        } catch(error) {
+            userData = await runQuery('SELECT * FROM users WHERE email=$1', [userName]);
+        } catch (error) {
             console.error(error);
         }
+        return userData;
     }
 
-    const deleteUser = async function(userID) {
+    const getUserByID = async function (userID) {
+        let userData = null;
         try {
-            await insertData(`DELETE FROM users WHERE userID=$1`, [userID]);
-        } catch(error) {
+            userData = await runQuery(`SELECT * FROM users WHERE userID=$1`, [userID]);
+        } catch (error) {
             console.error(error);
         }
+        return userData;
     }
 
-    const updateUser = async function(userID, userName, userEmail, userPassword) {
-        try {
-            await insertData(`UPDATE users SET name=$2, email=$3, password=$4 WHERE userID=$1`, [userID, userName, userEmail, userPassword]);
-        } catch(error) {
-            console.error(error);
+    const insertUser = async function (userName, userEmail, userPassword) {
+        let response = null;
+        if (await getUserByName(userName) || await getUserByEmail(userEmail)) {
+            response = DB_RESPONSES.ALREADY_EXIST;
+        } else {
+            try {
+                await insertData(`INSERT INTO users(name, email, password) VALUES ($1, $2, $3)`, [userName, userEmail, userPassword]);
+                response = DB_RESPONSES.OK;
+            } catch (error) {
+                console.error(error);
+            }
         }
+        return response;
+    }
+
+    const deleteUser = async function (userID) {
+        let response = null;
+        if (await getUserByID(userID)) {
+            try {
+                await insertData(`DELETE FROM users WHERE userID=$1`, [userID]);
+                response = DB_RESPONSES.OK;
+            } catch (error) {
+                console.error(error);
+            }
+        } else {
+            response = DB_RESPONSES.NOT_EXIST;
+        }
+        return response;
+    }
+
+    const updateUser = async function (userID, userName, userEmail, userPassword) {
+        let response = null;
+        let userToUpdate = await getUserByID(userID);
+        if (userToUpdate) {
+            let usernameCheck = await getUserByName(userName);
+            let useremailCheck = await getUserByEmail(userEmail);
+
+            if (usernameCheck && usernameCheck.name !== userToUpdate.name) {
+                response = DB_RESPONSES.ALREADY_EXIST;
+            } else if(useremailCheck && useremailCheck.email !== userToUpdate.email) {
+                response = DB_RESPONSES.ALREADY_EXIST;
+            } else {
+                try {
+                    await insertData(`UPDATE users SET name=$2, email=$3, password=$4 WHERE userID=$1`, [userID, userName, userEmail, userPassword]);
+                    response = DB_RESPONSES.OK;
+                } catch (error) {
+                    console.error(error);
+                }
+            }
+        } else {
+            response = DB_RESPONSES.NOT_EXIST;
+        }
+        return response;
     }
 
     return {
         getUser: getUserByID,
+        getUserByName: getUserByName,
         insertNewUser: insertUser,
         deleteExistingUser: deleteUser,
         updateExitingUser: updateUser
